pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "senti"
version = "0.1.0"
description = "Sentinel-Agent: a modular, security-first AI agent system"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "litellm>=1.40.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "docker>=7.0.0",
    "markdownify>=0.13.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/senti"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
